# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiMessage {
  error: string;
}

export type QuoteRequestInputService = typeof QuoteRequestInputService[keyof typeof QuoteRequestInputService];


export const QuoteRequestInputService = {
  nettoyage: 'nettoyage',
  polissage: 'polissage',
  ppf: 'ppf',
  covering: 'covering',
  autre: 'autre',
} as const;

export interface QuoteRequestInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  email: string;
  service: QuoteRequestInputService;
  /** @minLength 1 */
  message: string;
  marketingConsent?: boolean;
}

export interface QuoteRequest {
  id: number;
  name: string;
  phone: string;
  email: string;
  service: string;
  message: string;
  marketingConsent: boolean;
  createdAt: string;
}

export type RealisationService = typeof RealisationService[keyof typeof RealisationService];


export const RealisationService = {
  nettoyage: 'nettoyage',
  polissage: 'polissage',
  ppf: 'ppf',
  covering: 'covering',
} as const;

export interface Realisation {
  id: number;
  title: string;
  service: RealisationService;
  imageUrl: string;
  description: string;
  /** @nullable */
  vehicle?: string | null;
}

export interface ArticleSummary {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  coverImageUrl: string;
  publishedAt: string;
  /** Total number of times this article has been viewed */
  viewCount: number;
}

export interface Article {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  coverImageUrl: string;
  publishedAt: string;
  /** ISO timestamp of the last content modification (distinct from publishedAt) */
  updatedAt: string;
  /** Total number of times this article has been viewed */
  viewCount: number;
  content: string;
  related: ArticleSummary[];
}

/**
 * 'google' = avis Google vérifié, 'site' = témoignage recueilli directement
 */
export type TestimonialSource = typeof TestimonialSource[keyof typeof TestimonialSource];


export const TestimonialSource = {
  google: 'google',
  site: 'site',
} as const;

export interface Testimonial {
  id: number;
  name: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  /** @nullable */
  vehicle?: string | null;
  /** 'google' = avis Google vérifié, 'site' = témoignage recueilli directement */
  source: TestimonialSource;
}

export type ListRealisationsParams = {
service?: ListRealisationsService;
};

export type ListRealisationsService = typeof ListRealisationsService[keyof typeof ListRealisationsService];


export const ListRealisationsService = {
  nettoyage: 'nettoyage',
  polissage: 'polissage',
  ppf: 'ppf',
  covering: 'covering',
} as const;