# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a single article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ArticleSummary:
      type: object
      required: [id, slug, title, excerpt, coverImageUrl, publishedAt, viewCount]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        publishedAt:
          type: string
        viewCount:
          type: integer
          description: Total number of times this article has been viewed
    Article:
      type: object
      required: [id, slug, title, excerpt, coverImageUrl, publishedAt, updatedAt, viewCount, content, related]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        publishedAt:
          type: string
        updatedAt:
          type: string
          description: ISO timestamp of the last content modification (distinct from publishedAt)
        viewCount:
          type: integer
          description: Total number of times this article has been viewed
        content:
          type: string
        related:
          type: array
          items:
            $ref: "#/components/schemas/ArticleSummary"
...